% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task.R
\name{create_project}
\alias{create_project}
\title{OncoSimX Model Task}
\usage{
create_project(model, name, worksets = NULL)
}
\arguments{
\item{model}{Model name.}

\item{name}{Task name.}

\item{worksets}{Character vector of worksets.}
}
\description{
OncoSimX Model Task
}
