% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scenario.R
\name{create_scenario}
\alias{create_scenario}
\alias{create_workset}
\title{Manage Model Scenarios}
\usage{
create_scenario(model, name, base = NULL)

create_workset(model, name, base = NULL)
}
\arguments{
\item{model}{Model name or digest.}

\item{name}{New scenario name.}

\item{base}{Base run digest (optional).}
}
\value{
Nothing, invisibly.
}
\description{
Manage Model Scenarios
}
\examples{
if (FALSE) {
  create_scenario("RiskPaths", "NewScenario")
}


}
