% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task.R
\name{createTask}
\alias{createTask}
\title{Create new modeling task}
\usage{
createTask(dbCon, defRs, taskTxt = NA, setIds = NA)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}

\item{taskTxt}{Optional. Data frame specifying modeling task text:
\itemize{
\item \code{$name} — modeling task name (must be unique)
\item \code{$lang} — language code
\item \code{$descr} — modeling task description
\item \code{$note} — optional task notes
}}

\item{setIds}{Optional. Modeling task inputs: vector of integer workset IDs, or a data frame with a \code{$set_id} column.
Can supply a single workset ID or multiple IDs.}
}
\value{
Returns the ID of the new modeling task or 0L on error.
}
\description{
Create a new modeling task, which is a named set of multiple model inputs.
}
\details{
This function creates a new modeling task with optional text data (name, description, or notes)
and optional input working sets. You can create an empty task without any inputs or text and later update it
using \code{\link{updateTask}}.

Modeling task is a convenient way to bundle together multiple model inputs. After creation,
you can run the model by specifying the task name or task ID. The model will iterate through the task's
input worksets and produce output results for each.
}
\note{
To run examples you must have the modelOne database (\code{modelOne.sqlite}) in the current directory.
}
\examples{
myTaskTxt <- data.frame(
  name = "myTask",
  lang = c("EN", "FR"),
  descr = c("my first modeling task", "description in French"),
  note = c("this is a test task and includes two model input data sets with id 2 and 4", NA),
  stringsAsFactors = FALSE
)

theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

defRs <- getModel(theDb, "modelOne")

taskId <- createTask(theDb, defRs, myTaskTxt, c(2, 4))
if (taskId <= 0L) stop("task creation failed: ", defRs$modelDic$model_name, " ", defRs$modelDic$model_digest)

dbDisconnect(theDb)
# You can now run the model with the new task:
#   modelOne -OpenM.TaskName myTask

}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getTaskIdByName}}, \code{\link{updateTask}}
}
\author{
amc1999
}
\keyword{OpenM++}
\keyword{database}
