% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oms-UpdateModelRuns.R
\name{delete_model_run}
\alias{delete_model_run}
\alias{delete_run}
\alias{delete_model_runs}
\alias{delete_runs}
\title{Update Model Runs}
\usage{
delete_model_run(model, run)

delete_run(model, run)

delete_model_runs(model)

delete_runs(model)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}
}
\value{
A \code{list}, \code{tibble}, or nothing (invisibly).
}
\description{
Functions to update model runs. More information about these API endpoints
can be found at \href{https://github.com/openmpp/openmpp.github.io/wiki/Oms-web-service-API#update-model-runs}{here}.
}
\examples{
\dontrun{
use_OpenMpp_local()
delete_model_run("RiskPaths", "53300e8b56eabdf5e5fb112059e8c137")
}

}
