% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oms-ModelMetadata.R
\name{get_model}
\alias{get_model}
\alias{get_models}
\alias{get_models_list}
\title{Model Metadata}
\usage{
get_model(model)

get_models()

get_models_list()
}
\arguments{
\item{model}{Model digest or model name.}
}
\value{
A \code{list} or \code{tibble}.
}
\description{
Functions to get the list of models and load a specific model definition.
More information about these API endpoints can be found at
\href{https://github.com/openmpp/openmpp.github.io/wiki/Oms-web-service-API#get-model-metadata}{here}.
}
