% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_model_list}
\alias{get_model_list}
\alias{get_models}
\alias{get_model}
\title{Model}
\usage{
get_model_list()

get_models()

get_model(model)
}
\arguments{
\item{model}{Model digest or model name.}
}
\value{
A \code{list} or \code{tibble}.
}
\description{
Functions to get the complete list of models and load
a specific model.
}
