% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelTask.R
\name{get_model_task}
\alias{get_model_task}
\alias{get_model_tasks_list}
\alias{get_model_tasks}
\alias{get_model_task_worksets}
\alias{get_model_task_hist}
\alias{get_model_task_status}
\alias{get_model_task_run_list_status}
\alias{get_model_task_run_first}
\alias{get_model_task_run_last}
\alias{get_model_task_run_compl}
\alias{create_task}
\alias{update_task}
\alias{delete_task}
\title{Model Task}
\usage{
get_model_task(model, task)

get_model_tasks_list(model)

get_model_tasks(model)

get_model_task_worksets(model, task)

get_model_task_hist(model, task)

get_model_task_status(model, task, run)

get_model_task_run_list_status(model, task, run)

get_model_task_run_first(model, task)

get_model_task_run_last(model, task)

get_model_task_run_compl(model, task)

create_task(model, task, data)

update_task(model, task, data)

delete_task(model, task)
}
\arguments{
\item{model}{Model digest or model name.}

\item{task}{Modeling task.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}

\item{data}{Data used for the body of the POST request.}
}
\value{
A \code{list} from a JSON response object.
}
\description{
Functions for getting information about model tasks.
}
