% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parameters.R
\name{get_param_values}
\alias{get_param_values}
\alias{get_run_param_values}
\alias{get_workset_param_values}
\alias{get_model_run_param_values}
\alias{get_workset_param_values_csv}
\alias{get_run_param_values_csv}
\title{Workset Parameters}
\usage{
get_param_values(model, set, data)

get_run_param_values(model, run, data)

get_workset_param_values(model, set, name)

get_model_run_param_values(model, run, name, data)

get_workset_param_values_csv(model, set, name, data)

get_run_param_values_csv(model, run, name, data)
}
\arguments{
\item{model}{Model digest or model name.}

\item{set}{Name of workset (input set of model parameters).}

\item{data}{Data used for the body of the POST request.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}

\item{name}{Output table name.}
}
\value{
A \code{list} from a JSON response object.
}
\description{
Functions for retrieving parameters from worksets.
}
