% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Microdata.R
\name{get_run_micro_values}
\alias{get_run_micro_values}
\alias{get_model_mirco_value}
\alias{get_model_mirco_value_csv}
\title{Microdata}
\usage{
get_run_micro_values(model, run, data)

get_model_mirco_value(model, run, name, data)

get_model_mirco_value_csv(model, run, name, data)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}

\item{data}{Data used for the body of the POST request.}

\item{name}{Output table name.}
}
\value{
A \code{list} from a JSON response object.
}
\description{
Functions for retrieving parameters, output tables, and microdata from
models, model runs, and worksets.
}
