% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputTables.R
\name{get_run_table_values}
\alias{get_run_table_values}
\alias{get_run_table_calc}
\alias{get_model_table_expr}
\alias{get_model_table_expr_calc}
\alias{get_model_table_acc}
\alias{get_model_table_all_acc}
\alias{get_run_table_expr_csv}
\alias{get_run_table_calc_csv}
\alias{get_model_table_acc_csv}
\alias{get_model_table_all_acc_csv}
\title{Output Tables}
\usage{
get_run_table_values(model, run, data)

get_run_table_calc(model, run, data)

get_model_table_expr(model, run, name, data)

get_model_table_expr_calc(model, run, name, calc, data)

get_model_table_acc(model, run, name, data)

get_model_table_all_acc(model, run, name, data)

get_run_table_expr_csv(model, run, name)

get_run_table_calc_csv(model, run, name, calc)

get_model_table_acc_csv(model, run, name, data)

get_model_table_all_acc_csv(model, run, name, data)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}

\item{data}{Data used for the body of the POST request.}

\item{name}{Output table name.}

\item{calc}{Name of additional measure to calculate.}
}
\value{
A \code{list} from a JSON response object.
}
\description{
Functions for retrieving parameters, output tables, and microdata from
models, model runs, and worksets.
}
