% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Service.R
\name{get_service_config}
\alias{get_service_config}
\alias{get_service_state}
\alias{get_active_job_state}
\alias{get_queue_job_state}
\alias{get_hist_job_state}
\alias{set_queue_job_pos}
\alias{delete_job_hist}
\alias{get_archive_state}
\title{OpenM++ Service State}
\usage{
get_service_config()

get_service_state()

get_active_job_state(job)

get_queue_job_state(job)

get_hist_job_state(job)

set_queue_job_pos(pos, job)

delete_job_hist(job)

get_archive_state()
}
\arguments{
\item{job}{Model run submission time stamp.}

\item{pos}{Position.}
}
\value{
A \code{list} from a JSON response object, or nothing (invisibly).
}
\description{
Functions for retrieving or deleting service information.
}
