% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelWorkset.R
\name{get_workset}
\alias{get_workset}
\alias{get_worksets_list}
\alias{get_worksets}
\alias{get_workset_status}
\alias{get_workset_status_default}
\title{Model Worksets}
\usage{
get_workset(model, set)

get_worksets_list(model)

get_worksets(model)

get_workset_status(model, set)

get_workset_status_default(model, set)
}
\arguments{
\item{model}{Model digest or model name.}

\item{set}{Name of workset (input set of model parameters).}
}
\value{
A \code{list} from a JSON response object.
}
\description{
Functions for getting information about worksets.
}
