% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getDefaultWorksetId}
\alias{getDefaultWorksetId}
\title{Return default working set ID}
\usage{
getDefaultWorksetId(dbCon, defRs)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}
}
\value{
Positive integer (working set ID) on success. Stops execution if default workset is not found.
}
\description{
Returns the default working set ID for the model.
}
\details{
Returns the ID of the model's "default" working set of parameters.
Each model must have a default workset. The default workset is the first workset
of the model, i.e., the one with the minimum \code{set_id}.
Execution stops if no default workset is found.
}
\note{
To run examples you must have the \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model by name (if only one version exists)
defRs <- getModel(theDb, "modelOne")

# get default working set ID (expected to exist)
setId <- getDefaultWorksetId(theDb, defRs)

# try to find "myData" working set ID
setId <- getWorksetIdByName(theDb, defRs, "myData")
if (setId <= 0L) warning("workset not found: ", "myData")

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getFirstRunId}}, \code{\link{getLastRunId}},
\code{\link{getWorksetIdByName}}, \code{\link{getWorksetRunIds}}
}
\author{
amc1999
}
