% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getFirstRunId}
\alias{getFirstRunId}
\title{Return ID of first model run results}
\usage{
getFirstRunId(dbCon, defRs)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}
}
\value{
Run ID: positive integer on success or negative if not found.
}
\description{
Returns the ID of the first model run results for a given model.
}
\details{
Returns the ID of the first model run results. This is a positive integer.
If the model does not have any run results (the model was never executed), the function returns a negative value.
}
\note{
To run examples you must have the \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model by name (use this if only one version of the model exists)
defRs <- getModel(theDb, "modelOne")

# get first run ID of the model
runId <- getFirstRunId(theDb, defRs)
if (runId <= 0L) warning("model run results not found")

dbDisconnect(theDb)

}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getLastRunId}}, \code{\link{getModel}}, \code{\link{getTaskIdByName}}
}
\author{
amc1999
}
\keyword{OpenM++}
\keyword{database}
