% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getLanguages}
\alias{getLanguages}
\title{Return list of languages}
\usage{
getLanguages(dbCon)
}
\arguments{
\item{dbCon}{Database connection.}
}
\value{
A data frame with columns:
\describe{
\item{lang_id}{Language ID.}
\item{lang_code}{Language code.}
\item{lang_name}{Language name.}
}
}
\description{
Returns the list of languages from the \code{lang_lst} table in the database.
}
\details{
Reads the language list from the database and returns it as a data frame.
}
\note{
To run examples you must have the \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

langRs <- getLanguages(theDb)

dbDisconnect(theDb)

}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}
}
\author{
amc1999
}
\keyword{OpenM++}
\keyword{database}
