% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getModel.R
\name{getModel}
\alias{getModel}
\title{Find OpenM++ model definition in database}
\usage{
getModel(dbCon, modelName, modelDigest = NA)
}
\arguments{
\item{dbCon}{Database connection.}

\item{modelName}{Model name, e.g., \code{"modelOne"}.}

\item{modelDigest}{Optional model digest string, e.g., \code{"_201208171604590148_"}.
If \code{NA} or \code{NULL}, the first model with minimum \code{model_id} is used.}
}
\value{
A list of database rows describing the model, including:
\describe{
\item{modelDic}{\code{model_dic} row}
\item{langLst}{list of languages (\code{lang_lst} table rows)}
\item{typeDic}{\code{type_dic} rows of the model}
\item{typeEnum}{\code{type_enum_lst} rows of the model, optional, can be empty}
\item{paramDic}{\code{parameter_dic} rows of the model}
\item{paramDims}{\code{parameter_dims} rows of the model, optional, can be empty}
\item{tableDic}{\code{table_dic} rows of the model}
\item{tableDims}{\code{table_dims} rows of the model, optional, can be empty}
\item{tableAcc}{\code{table_acc} rows of the model}
\item{tableExpr}{\code{table_expr} rows of the model}
}
}
\description{
Find a model in the database and return its OpenM++ model definition, including input parameters and output tables.
}
\details{
Finds a model in the database and returns its OpenM++ model definition, including input parameters and output tables.

Models can be identified by name and digest string (serving as the version).
If only one version of a model exists in the database, the \code{modelDigest} argument can be omitted.
If multiple models with the same name exist and \code{modelDigest} is \code{NA} or \code{NULL}, the model with the minimum \code{model_id} is returned.
}
\note{
To run examples, you must have the \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# Get model by name (single version)
defRs <- getModel(theDb, "modelOne")

# Get model by name and digest (specific version)
def2012August17 <- getModel(theDb, "modelOne", "_201208171604590148_")

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getLanguages}}
}
\author{
amc1999
}
