% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTaskIdByName}
\alias{getTaskIdByName}
\title{Return modeling task ID by name}
\usage{
getTaskIdByName(dbCon, defRs = NA, taskName)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Optional model definition: database rows describing model input parameters and output tables. Default is NA.}

\item{taskName}{Name of the modeling task.}
}
\value{
Modeling task ID: positive integer on success or negative if not found.
}
\description{
Returns the ID of a modeling task given its name.
}
\details{
Returns the ID of the modeling task matching the specified name.
If no task exists with that name, a negative value is returned.
If multiple tasks have the same name, the minimum task ID is returned.

If the \code{defRs} argument is supplied, the search is restricted to the specific model.

A modeling task is a named set of model inputs that contains a name and a vector of model workset IDs.
See \code{\link{createWorkset}} for details about worksets (working sets of model input parameters).

Modeling tasks are a convenient way to bundle multiple inputs of the model.
Once a task is created, the model can be run by specifying the task name or task ID.
The model will iterate through the task's input worksets and produce output results for each input.
}
\note{
To run examples you must have the \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# try to find "taskOne" task ID for any model
taskId <- getTaskIdByName(theDb, taskName = "taskOne")
if (taskId <= 0L) warning("task not found: ", "taskOne")

# get model by name: use this if only one version of the model exists
defRs <- getModel(theDb, "modelOne")

# try to find "taskOne" task ID for "modelOne"
taskId <- getTaskIdByName(theDb, defRs, "taskOne")
if (taskId <= 0L) warning("task not found: ", "taskOne")

dbDisconnect(theDb)

}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{createTask}}, \code{\link{selectTask}}, \code{\link{selectTaskRun}}
}
\author{
amc1999
}
\keyword{OpenM++}
\keyword{database}
