% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getTaskLastRunId}
\alias{getTaskLastRunId}
\title{Return last ID of modeling task run}
\usage{
getTaskLastRunId(dbCon, taskId)
}
\arguments{
\item{dbCon}{Database connection.}

\item{taskId}{Modeling task ID.}
}
\value{
Task run ID: positive integer on success or negative if not found.
}
\description{
Returns the ID of the last (most recent) run for a specified modeling task.
}
\details{
Returns the ID of the most recent run associated with the given modeling task.
If there are no runs for the specified task, a negative value is returned.
}
\note{
To run examples you must have the \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model by name (if only one version exists)
defRs <- getModel(theDb, "modelOne")

# get task ID of "taskOne"
taskId <- getTaskIdByName(theDb, defRs, "taskOne")
if (taskId <= 0L) warning("task not found: taskOne")

# get last task run ID
lastId <- getTaskLastRunId(theDb, taskId)
if (lastId <= 0L) warning("task run(s) not found")

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getTaskFirstRunId}}, \code{\link{getModel}}, \code{\link{getTaskIdByName}}
}
\author{
amc1999
}
