% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getWorksetIdByName}
\alias{getWorksetIdByName}
\title{Return working set ID by name}
\usage{
getWorksetIdByName(dbCon, defRs, worksetName)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}

\item{worksetName}{Name of the parameters working set.}
}
\value{
Positive integer on success (working set ID), or negative if not found.
}
\description{
Returns the ID of a working set given its name.
}
\details{
Returns the working set ID for the specified name.
If the model has no working set with that name, a negative value is returned.
If the model has multiple working sets with the same name, the minimum set ID is returned.
}
\note{
To run examples you must have the \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model by name (if only one version exists)
defRs <- getModel(theDb, "modelOne")

# get default working set ID by name (expected to exist)
setId <- getWorksetIdByName(theDb, defRs, defRs$modelDic$model_name)

# try to find "myData" working set ID
setId <- getWorksetIdByName(theDb, defRs, "myData")
if (setId <= 0L) warning("workset not found: ", "myData")

dbDisconnect(theDb)

}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getDefaultWorksetId}}, \code{\link{getWorksetRunIds}}
}
\author{
amc1999
}
\keyword{OpenM++}
\keyword{database}
