% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getWorksetRunIds}
\alias{getWorksetRunIds}
\title{Return IDs of model run results for a specified working set}
\usage{
getWorksetRunIds(dbCon, worksetId)
}
\arguments{
\item{dbCon}{Database connection.}

\item{worksetId}{ID of parameters working set (must be a positive integer).}
}
\value{
Data frame with an integer column \code{$run_id}.
}
\description{
Returns IDs of model run results where input parameters are from a specified working set.
}
\details{
Returns IDs of model run results where input parameters are from the specified working set.

Note: There is no established link in the database between an input data working set and model run results.
Input data can be modified or deleted after a model run. If you need the input parameter values for a specific run,
use \code{\link{selectRunParameter}} instead.

It is recommended to create a modeling task using \code{\link{createTask}}
to include multiple working sets of input parameters and run the model using these inputs.
}
\note{
To run examples you must have the \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

runIdRs <- getWorksetRunIds(theDb, 2L)

dbDisconnect(theDb)

}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{createTask}}, \code{\link{getModel}}, \code{\link{getFirstRunId}},
\code{\link{getLastRunId}}, \code{\link{getDefaultWorksetId}},
\code{\link{getWorksetIdByName}}, \code{\link{selectRunParameter}}
}
\author{
amc1999
}
\keyword{OpenM++}
\keyword{database}
