% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskInternal.R
\name{insertTaskInputWorksetsInternal}
\alias{insertTaskInputWorksetsInternal}
\alias{updateTaskSetIds}
\title{Internal: Insert workset IDs as task input}
\usage{
updateTaskSetIds(dbCon, i_taskId, i_setIds = NA)
}
\arguments{
\item{dbCon}{Database connection.}

\item{i_taskId}{Modeling task ID.}

\item{i_setIds}{Optional workset IDs to add to the task input. Can be:
\describe{
\item{scalar}{Single positive integer workset ID.}
\item{vector}{Vector of positive integer workset IDs.}
\item{data frame}{Data frame with a \code{\$set_id} column of positive integers.}
}
All workset IDs must exist in \code{workset_lst\$set_id} table.}
}
\value{
Returns TRUE on success, FALSE otherwise.
}
\description{
For internal use only. Insert additional workset IDs as inputs into the \code{task_set} table.
Must be executed within a transaction scope.
}
\details{
This function inserts one or more workset IDs into the \code{task_set} table for a given task.
It does not update or remove existing task inputs; it only adds new ones.
Must be called inside a transaction to ensure database consistency.
}
\note{
Internal function: not intended for direct user calls.
Higher-level functions like \code{createTask} or \code{updateTask} use this internally.
}
