% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{is_compatible}
\alias{is_compatible}
\title{Are data frames compatible?}
\usage{
is_compatible(x, y)
}
\arguments{
\item{x}{New parameters.}

\item{y}{Old parameters}
}
\description{
Are data frames compatible?
}
