% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModel.R
\name{load_model}
\alias{load_model}
\alias{OncoSimXModel}
\title{OncoSimX Model Class}
\usage{
load_model(model, ...)
}
\arguments{
\item{model}{Model digest or model name.}

\item{...}{Not currently used.}
}
\description{
OncoSimX Model Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{digest}}{Model digest.}

\item{\code{name}}{Model name.}

\item{\code{version}}{Model version.}

\item{\code{params}}{Character vector of model parameters.}

\item{\code{tables}}{Character vector of output tables.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{digest}}{Model digest.}

\item{\code{name}}{Model name.}

\item{\code{version}}{Model version.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXModel-new}{\code{OncoSimXModel$new()}}
\item \href{#method-OncoSimXModel-print}{\code{OncoSimXModel$print()}}
\item \href{#method-OncoSimXModel-clone}{\code{OncoSimXModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModel-new}{}}}
\subsection{Method \code{new()}}{
Create a new OncoSimXModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModel$new(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model digest or name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{OncoSimXModel} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModel-print}{}}}
\subsection{Method \code{print()}}{
Print a OncoSimXModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not currently used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
