% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModelRun.R
\name{load_model_run}
\alias{load_model_run}
\alias{OncoSimXModelRun}
\title{OncoSimX Model Run Class}
\usage{
load_model_run(model, run)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}
}
\description{
OncoSimX Model Run Class
}
\section{Super class}{
\code{\link[oncosim:OncoSimXModel]{oncosim::OncoSimXModel}} -> \code{OncoSimXModelRun}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXModelRun-new}{\code{OncoSimXModelRun$new()}}
\item \href{#method-OncoSimXModelRun-clone}{\code{OncoSimXModelRun$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="oncosim" data-topic="OncoSimXModel" data-id="print"><a href='../../oncosim/html/OncoSimXModel.html#method-OncoSimXModel-print'><code>oncosim::OncoSimXModel$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$new(model, run)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
