% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModelTask.R
\name{load_model_task}
\alias{load_model_task}
\title{OncoSimX Model Task Class}
\usage{
load_model_task(model, task)
}
\arguments{
\item{model}{Model digest or model name.}

\item{task}{Modeling task name.}
}
\value{
An \code{OncoSimXModelTask} instance.
}
\description{
OncoSimX Model Task Class
}
