% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncosim-package.R
\docType{package}
\name{oncosim-package}
\alias{oncosim}
\alias{oncosim-package}
\title{oncosim: Programmatic interface to the OncoSimX Web Platform}
\description{
This package provides a programmatic interfact to the OncoSimX web-based platform. OncoSim is a free, web-based simulation tool that can be used to evaluate cancer control strategies.
}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{warkentin@lunenfeld.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID})

}
\keyword{internal}
