% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunOptions.R
\name{opts_run}
\alias{opts_run}
\alias{print.OpenMRunOpts}
\title{Run Configuration Options}
\usage{
opts_run(
  SimulationCases = "5000",
  SubValues = "12",
  RunStamp = NULL,
  Tables = list(),
  ...
)

\method{print}{OpenMRunOpts}(x, ...)
}
\arguments{
\item{SimulationCases}{Number of cases to simulate.}

\item{SubValues}{Number of sub values.}

\item{RunStamp}{Run stamp.}

\item{Tables}{List of tables to output.}

\item{...}{Not currently used.}

\item{x}{Object to print.}
}
\description{
Run Configuration Options
}
