% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunModels.R
\name{run_model}
\alias{run_model}
\alias{get_model_run_state}
\alias{get_run_state}
\alias{stop_model_run}
\alias{stop_run}
\title{Run Models and Monitor Progress}
\usage{
run_model(data)

get_model_run_state(model, stamp)

get_run_state(model, stamp)

stop_model_run(model, stamp)

stop_run(model, stamp)
}
\arguments{
\item{data}{Data used for the body of the POST request.}

\item{model}{Model digest or name.}

\item{stamp}{Model run stamp.}
}
\value{
A \code{list} or nothing, invisibly.
}
\description{
Run Models and Monitor Progress
}
