% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRunAccumulator.R
\name{selectRunAccumulator}
\alias{selectRunAccumulator}
\title{Select output table accumulator values from model run result}
\usage{
selectRunAccumulator(dbCon, defRs, runId, tableName, accName = NA)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}

\item{runId}{Positive integer. ID of model run results.}

\item{tableName}{Output table name (character), e.g., \code{"sexAge"}.}

\item{accName}{Optional output table accumulator name (character), e.g., \code{"acc0"}.
If missing or \code{NA}, all accumulators are returned.}
}
\value{
Data frame of database rows with output table accumulator(s) dimensions and values:
\describe{
\item{acc_id}{Accumulator number.}
\item{sub_id}{Sub-value number.}
\item{dim0,...,dimN}{Dimension items enum IDs (not returned if rank is zero).}
\item{acc_value}{Accumulator value.}
}
}
\description{
Retrieve output table accumulator values from a specific model run result.
}
\details{
Returns values of output table accumulators from the model run results with the specified \code{runId}.
You can get either a single accumulator by specifying \code{accName} or all accumulators if it is missing.

The \code{defRs} object must be obtained via \verb{\code{\link{getModel}}}.
}
\note{
To run examples you must have \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model definition
defRs <- getModel(theDb, "modelOne")

# get first run id
runId <- getFirstRunId(theDb, defRs)
if (runId <= 0L) stop("model run results not found")

# select accumulator(s) value of "salarySex"
allAccValueRs <- selectRunAccumulator(theDb, defRs, runId, "salarySex")
acc0_ValueRs <- selectRunAccumulator(theDb, defRs, runId, "salarySex", "acc0")

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getFirstRunId}}, \code{\link{getLastRunId}},
\code{\link{selectRunOutputValue}}, \code{\link{selectRunParameter}}
}
\author{
amc1999
}
