% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRunValue.R
\name{selectRunOutputValue}
\alias{selectRunOutputValue}
\title{Select output table expression values from model run result}
\usage{
selectRunOutputValue(dbCon, defRs, runId, tableName, exprName = NA)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}

\item{runId}{Positive integer. ID of model run results.}

\item{tableName}{Output table name (character), e.g., \code{"sexAge"}.}

\item{exprName}{Optional output table expression name (character), e.g., \code{"AverageAge"}.
If missing or \code{NA}, returns all expressions.}
}
\value{
Data frame of database rows with output table expression(s) dimensions and values:
\describe{
\item{expr_id}{Expression number (zero-based).}
\item{dim0,...,dimN}{Dimension items enum IDs (not returned if rank is zero).}
\item{value}{Output table expression value.}
}
}
\description{
nRetrieve output table expression values from a specific model run result.
}
\details{
Returns values of output table expressions from the model run results with the specified \code{runId}.
You can retrieve either a single expression by specifying its name or all expressions if
\code{exprName} is missing. The \code{defRs} object must be obtained via \verb{\code{\link{getModel}}}.
}
\note{
To run examples you must have \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model definition
defRs <- getModel(theDb, "modelOne")

# get first run id
runId <- getFirstRunId(theDb, defRs)
if (runId <= 0L) stop("model run results not found")

# select output table expression(s) value
expr2_ValueRs <- selectRunOutputValue(theDb, defRs, runId, "salarySex", "expr2")
allExprValueRs <- selectRunOutputValue(theDb, defRs, runId, "salarySex")

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getFirstRunId}}, \code{\link{getLastRunId}},
\code{\link{selectRunAccumulator}}, \code{\link{selectRunParameter}}
}
\author{
amc1999
}
