% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRunParameter.R
\name{selectRunParameter}
\alias{selectRunParameter}
\title{Select parameter values from model run result}
\usage{
selectRunParameter(dbCon, defRs, runId, paramName)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}

\item{runId}{Positive integer. ID of model run results.}

\item{paramName}{Parameter name (character), e.g., \code{"sexAge"}.}
}
\value{
Data frame of database rows with parameter dimensions and values:
\describe{
\item{sub_id}{Parameter sub-value ID or zero if parameter has no sub-values.}
\item{dim0,...,dimN}{Dimension items enum IDs (not returned if rank is zero).}
\item{value}{Parameter value.}
}
}
\description{
Retrieve input parameter values from a specific model run result.
}
\details{
Returns values of an input parameter from the model run results with the specified \code{runId}.
The \code{defRs} object must be obtained via \verb{\code{\link{getModel}}}.
}
\note{
To run examples you must have \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model definition
defRs <- getModel(theDb, "modelOne")

# get first run id
runId <- getFirstRunId(theDb, defRs)
if (runId <= 0L) stop("model run results not found")

# select parameter "ageSex" value
paramValueRs <- selectRunParameter(theDb, defRs, runId, "ageSex")

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getFirstRunId}}, \code{\link{getLastRunId}},
\code{\link{selectRunAccumulator}}, \code{\link{selectRunOutputValue}}
}
\author{
amc1999
}
