% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTask.R
\name{selectTask}
\alias{selectTask}
\title{Select modeling task}
\usage{
selectTask(dbCon, taskId)
}
\arguments{
\item{dbCon}{Database connection.}

\item{taskId}{Modeling task ID (positive integer).}
}
\value{
A list of database rows:
\describe{
\item{taskLst}{task_lst row: task ID and name.}
\item{taskTxt}{task_txt rows: language, description, notes.}
\item{taskSet}{task_set rows: task input workset IDs.}
}
}
\description{
Retrieve modeling task text (name, description, notes) and input workset IDs.
}
\details{
Returns the text and input worksets of a modeling task:
\itemize{
\item Task text includes the task name, description, and optional notes in multiple languages.
\item Task input worksets are the list of working set IDs associated with the task.
}
}
\note{
To run examples you must have \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model definition
defRs <- getModel(theDb, "modelOne")

# find task ID
taskId <- getTaskIdByName(theDb, defRs, "taskOne")
if (taskId <= 0L) stop("task: ", "taskOne", " not found for model: ",
                       defRs$modelDic$model_name, " ", defRs$modelDic$model_digest)

# select task text and input worksets
taskRs <- selectTask(theDb, taskId)

# select list of all tasks for the model
taskLstRs <- selectTaskList(theDb, defRs)

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getTaskIdByName}}, \code{\link{getTaskFirstRunId}},
\code{\link{getTaskLastRunId}}, \code{\link{selectTaskList}},
\code{\link{selectTaskRun}}, \code{\link{selectTaskRunList}}
}
\author{
amc1999
}
