% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTask.R
\name{selectTaskList}
\alias{selectTaskList}
\title{Select list of modeling tasks}
\usage{
selectTaskList(dbCon, defRs)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}
}
\value{
A list of database rows:
\describe{
\item{taskLst}{task_lst rows: task ID and name.}
\item{taskTxt}{task_txt rows: language, description, notes.}
}
}
\description{
Retrieve the list of modeling tasks for a specified model.
}
\details{
Returns the list of modeling tasks for the specified model. Each task includes its text (name, description, notes)
and ID. Useful for enumerating all tasks available for a given model.
}
\note{
To run examples you must have \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model definition
defRs <- getModel(theDb, "modelOne")

# select list of all tasks for the model
taskLstRs <- selectTaskList(theDb, defRs)

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getTaskIdByName}}, \code{\link{getTaskFirstRunId}},
\code{\link{getTaskLastRunId}}, \code{\link{selectTask}}, \code{\link{selectTaskRun}},
\code{\link{selectTaskRunList}}
}
\author{
amc1999
}
