% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectTaskRun.R
\name{selectTaskRunList}
\alias{selectTaskRunList}
\title{Select list of modeling task runs}
\usage{
selectTaskRunList(dbCon, taskId)
}
\arguments{
\item{dbCon}{Database connection.}

\item{taskId}{ID of the modeling task.}
}
\value{
A list of database rows:
\describe{
\item{taskLst}{task_lst row: task ID and name.}
\item{taskRunLst}{task_run_lst rows: task run ID, status, and time.}
}
}
\description{
Retrieve the list of runs for a specific modeling task, including task run ID, status, and time.
}
\details{
Returns for the specified task ID a list of task run IDs, their status, and timestamps.

Task run status can be one of:
\describe{
\item{i}{Not yet started.}
\item{p}{Run in progress.}
\item{w}{Run in progress, waiting for additional input.}
\item{s}{Completed successfully.}
\item{e}{Failed (error).}
\item{x}{Reserved.}
}

Status \code{w} indicates that the task can be dynamically updated by an external script.
The model executable is waiting for additional input or a "ready to be completed" signal.
Use \code{\link{setTaskWaitCompleted}} to mark the task as ready to be completed,
and \code{\link{updateTask}} to insert additional input.
}
\note{
To run examples you must have \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model definition
defRs <- getModel(theDb, "modelOne")

# get task id of "taskOne"
taskId <- getTaskIdByName(theDb, defRs, "taskOne")
if (taskId <= 0L) stop("task not found: taskOne")

# last (most recent) run of that task
lastId <- getTaskLastRunId(theDb, taskId)
if (lastId <= 0L) stop("task run(s) not found")

# select task run status, input worksets, and output results
taskRunRs <- selectTaskRun(theDb, lastId)

# select list of all task runs for the task
taskRunLstRs <- selectTaskRunList(theDb, taskId)

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getTaskIdByName}}, \code{\link{getTaskFirstRunId}},
\code{\link{getTaskLastRunId}}, \code{\link{selectTask}}, \code{\link{selectTaskList}},
\code{\link{selectTaskRun}}, \code{\link{setTaskWaitCompleted}}
}
\author{
amc1999
}
