% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setReadonlyWorkset.R
\name{setReadonlyDefaultWorkset}
\alias{setReadonlyDefaultWorkset}
\title{Set or clear read-only status for default working set of model parameters}
\usage{
setReadonlyDefaultWorkset(dbCon, defRs, isReadonly)
}
\arguments{
\item{dbCon}{Database connection.}

\item{defRs}{Model definition: database rows describing model input parameters and output tables.}

\item{isReadonly}{Logical; if \code{TRUE}, mark the default working set as read-only, otherwise clear read-only status.}
}
\value{
Returns the working set ID of the default workset, or 0L if no worksets exist.
}
\description{
Mark the default working set of model parameters as read-only or clear its read-only status.
}
\details{
A workset must be \strong{not read-only} to allow parameter updates using \code{\link{updateWorksetParameter}}.
Conversely, a workset must be \strong{read-only} to run the model.
Typically, you wrap \code{\link{updateWorksetParameter}} calls with \code{setReadonlyDefaultWorkset}
or \code{\link{setReadonlyWorkset}} to temporarily allow updates.

A workset can be a full set containing all model parameters or a subset containing only some parameters.
Each model must have a "default" workset: the first workset of the model with \code{set_id = min(set_id)}, which always contains all parameters.

You must use \code{\link{getModel}} to obtain the model definition \code{defRs}.
}
\note{
To run examples you must have \code{modelOne} database \code{modelOne.sqlite} in the current directory.
}
\examples{
theDb <- dbConnect(RSQLite::SQLite(), "modelOne.sqlite", synchronous = "full")
invisible(dbGetQuery(theDb, "PRAGMA busy_timeout = 86400")) # recommended

# get model definition
defRs <- getModel(theDb, "modelOne")

# reset read-only status of default workset
setId <- setReadonlyDefaultWorkset(theDb, defRs, FALSE)
if (setId <= 0L) stop("no worksets exist for model: ", defRs$modelDic$model_name, " ", defRs$modelDic$model_digest)

# you can update model parameters now
# updateWorksetParameter(theDb, defRs, setId, ageSex, salaryAge)

# make default workset read-only to run the model
setReadonlyDefaultWorkset(theDb, defRs, TRUE)

dbDisconnect(theDb)
}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\code{\link{getModel}}, \code{\link{getDefaultWorksetId}}, \code{\link{getWorksetIdByName}},
\code{\link{copyWorksetParameterFromRun}}, \code{\link{setReadonlyWorkset}},
\code{\link{updateWorksetParameter}}
}
\author{
amc1999
}
