% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{toQuoted}
\alias{toQuoted}
\title{Return SQL-quoted string}
\usage{
toQuoted(srcStr)
}
\arguments{
\item{srcStr}{Source string to be SQL-quoted.}
}
\value{
SQL-quoted string suitable for use in SQL statements.
}
\description{
Return a SQL-quoted string. For example, \verb{O'Connor} becomes \verb{'O''Connor'}.
}
\details{
This function duplicates each apostrophe in the input string and encloses the result in single quotes.
For example, \verb{O'Connor} becomes \verb{'O''Connor'}.
}
\examples{
someName <- toQuoted("O'Connor")
# someName
# [1] "'O''Connor'"

sql <- paste(
  "SELECT * FROM someTable WHERE name = ", toQuoted("O'Connor"), sep=""
)
# sql
# [1] "SELECT * FROM someTable WHERE name = 'O''Connor'"

}
\references{
OpenM++ documentation: \url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\seealso{
\url{https://github.com/openmpp/openmpp.github.io/wiki}
}
\keyword{OpenM++}
\keyword{database}
