% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelProfile.R
\name{touch_model_profile}
\alias{touch_model_profile}
\alias{delete_model_profile}
\alias{set_model_profile_opt}
\alias{delete_model_profile_opt}
\title{OncoSimX Model Profile}
\usage{
touch_model_profile(model, data)

delete_model_profile(model, profile)

set_model_profile_opt(model, profile, key, value)

delete_model_profile_opt(model, profile, key)
}
\arguments{
\item{model}{Model digest or model name.}

\item{data}{Data used for the body of the POST request.}

\item{profile}{Profile name.}

\item{key}{Option key.}

\item{value}{Option value.}
}
\value{
A \code{list} from a JSON response object or nothing (invisibly).
}
\description{
Functions for creating, modifying, and deleting profiles and profile options.
}
