% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worksetParameterInternal.R
\name{updateWorksetParameterTxt}
\alias{updateWorksetParameterTxt}
\title{Insert or update workset parameter notes}
\usage{
updateWorksetParameterTxt(dbCon, i_paramDef, i_wsParamTxt = NULL)
}
\arguments{
\item{dbCon}{Database connection.}

\item{i_paramDef}{List identifying parameter in a workset:
\describe{
\item{\$setId}{Workset ID.}
\item{\$paramHid}{Parameter HID.}
}}

\item{i_wsParamTxt}{Workset parameter text (notes):
Data frame with columns:
\describe{
\item{\$lang}{Language code.}
\item{\$note}{Parameter notes.}
}}
}
\value{
Invisibly returns number of rows inserted/updated.
}
\description{
Internal use only. Inserts or updates notes for a specific parameter in a workset.
Must be called within a transaction scope.
}
\details{
This function updates the \code{workset_parameter_txt} table.
If notes for the given parameter and language exist, they are updated;
otherwise, a new row is inserted.

Must be executed within a transaction scope to ensure consistency.
}
\author{
amc1999
}
\keyword{internal}
