% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetadata.R
\name{get_model}
\alias{get_model}
\alias{get_models}
\alias{get_models_list}
\title{OpenM++ Model Metadata}
\usage{
get_model(model)

get_models()

get_models_list()
}
\arguments{
\item{model}{Model digest or model name.}
}
\value{
A \code{list} or \code{tibble}.
}
\description{
Functions to get the list of models and load a specific model definition.
}
