% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{get_model}
\alias{get_model}
\alias{get_models}
\alias{get_models_list}
\alias{run_model}
\alias{stop_model_run}
\title{OncoSimX Models}
\usage{
get_model(model)

get_models()

get_models_list()

run_model(data)

stop_model_run(model, stamp)
}
\arguments{
\item{model}{Model digest or model name.}

\item{data}{Data used for the body of the POST request.}

\item{stamp}{Model run stamp or modeling task run stamp.}
}
\value{
A \code{list} or \code{tibble}.
}
\description{
Functions to get the complete list of models and load, run, or stop
a specific model.
}
