% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelRun.R
\name{get_model_run}
\alias{get_model_run}
\alias{get_model_runs_list}
\alias{get_model_runs}
\alias{get_runs}
\alias{get_model_run_status}
\alias{get_run_status}
\alias{get_model_run_list_status}
\alias{get_model_run_status_first}
\alias{get_model_run_status_last}
\alias{get_model_run_status_compl}
\alias{delete_model_run}
\alias{delete_run}
\title{OpenM++ Model Run}
\usage{
get_model_run(model, run)

get_model_runs_list(model)

get_model_runs(model)

get_runs(model)

get_model_run_status(model, run)

get_run_status(model, run)

get_model_run_list_status(model, run)

get_model_run_status_first(model)

get_model_run_status_last(model)

get_model_run_status_compl(model)

delete_model_run(model, run)

delete_run(model, run)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}
}
\value{
A \code{list}, \code{tibble}, or nothing (invisibly).
}
\description{
Functions to retrieve and delete model runs.
}
