% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutputTables.R
\name{get_run_table}
\alias{get_run_table}
\alias{get_run_table_csv}
\title{OpenM++ Output Tables}
\usage{
get_run_table(model, run, name)

get_run_table_csv(model, run, name)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}

\item{name}{Output table name.}
}
\value{
A \code{list} or \code{tibble}.
}
\description{
Functions for retrieving output tables from model runs.
}
