% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oms-UserSettings.R
\name{get_user_views}
\alias{get_user_views}
\alias{set_user_views}
\alias{delete_user_views}
\title{Manage OpenM++ User Settings}
\usage{
get_user_views(model)

set_user_views(model, data)

delete_user_views(model)
}
\arguments{
\item{model}{Model digest or model name.}

\item{data}{Data used for the body of the request.}
}
\value{
A \code{list} from JSON response object or nothing (invisibly).
}
\description{
Functions for getting, setting, and deleting user settings. More information
about these API endpoints can be found at \href{https://github.com/openmpp/openmpp.github.io/wiki/Oms-web-service-API#user-manage-user-settings}{here}.
}
\examples{
\dontrun{
use_OpenMpp_local()
get_user_views("RiskPaths")
}

}
