% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/User.R
\name{get_user_views}
\alias{get_user_views}
\alias{set_user_views}
\alias{del_user_views}
\title{Manage OpenM++ User Settings}
\usage{
get_user_views(model)

set_user_views(model, data)

del_user_views(model)
}
\arguments{
\item{model}{Model digest or model name.}

\item{data}{Data used for the body of the POST request.}
}
\value{
A \code{list} from JSON response object or nothing (invisibly).
}
\description{
Functions for getting, setting, and deleting user configuration settings.
}
