% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelWorkset.R, R/Scenario.R
\name{get_workset}
\alias{get_workset}
\alias{get_worksets_list}
\alias{get_worksets}
\alias{get_scenarios}
\alias{get_workset_status}
\alias{get_workset_status_default}
\alias{set_workset_readonly}
\alias{create_workset}
\alias{merge_workset}
\alias{update_workset_param_csv}
\alias{replace_workset}
\alias{delete_workset}
\alias{delete_workset_param}
\alias{update_workset_param}
\alias{copy_param_run_to_workset}
\alias{merge_param_run_to_workset}
\alias{copy_param_workset_to_workset}
\alias{merge_param_workset_to_workset}
\alias{upload_workset_params}
\alias{delete_scenario}
\title{OpenM++ Model Worksets}
\usage{
get_workset(model, set)

get_worksets_list(model)

get_worksets(model)

get_scenarios(model)

get_workset_status(model, set)

get_workset_status_default(model)

set_workset_readonly(model, set, readonly)

create_workset(data)

merge_workset(workset)

update_workset_param_csv(workset, csv)

replace_workset(workset)

delete_workset(model, set)

delete_workset_param(model, set, name)

update_workset_param(model, set, name, data)

copy_param_run_to_workset(model, set, name, run)

merge_param_run_to_workset(model, set, name, run)

copy_param_workset_to_workset(model, set, name, from)

merge_param_workset_to_workset(model, set, name, from)

upload_workset_params(model, set, data)

delete_scenario(model, set)
}
\arguments{
\item{model}{Model digest or model name.}

\item{set}{Name of workset (input set of model parameters).}

\item{readonly}{Boolean. Should workset be read-only?}

\item{data}{Data used for the body of the request.}

\item{workset}{Workset metadata.}

\item{csv}{CSV file path.}

\item{name}{Output table name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}

\item{from}{Source workset name.}
}
\value{
A \code{list}, \code{tibble}, or nothing (invisibly).
}
\description{
Functions for creating, copying, merging, retrieving, and deleting
worksets.
}
