% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Uploads.R
\name{initiate_run_upload}
\alias{initiate_run_upload}
\alias{initiate_workset_upload}
\alias{delete_upload_files}
\title{Upload model run or workset}
\usage{
initiate_run_upload(model, run, data)

initiate_workset_upload(model, set, data)

delete_upload_files(folder)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}

\item{data}{Data used for the body of the POST request.}

\item{set}{Name of workset (input set of model parameters).}

\item{folder}{Download folder file name.}
}
\value{
A \code{list} or \code{tibble}.
}
\description{
Functions to download model, model run results, or input parameters.
}
