% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModel.R
\name{load_model}
\alias{load_model}
\alias{OncoSimXModel}
\title{OncoSimX Model Class}
\usage{
load_model(model)
}
\arguments{
\item{model}{Model digest or model name.}
}
\value{
An \code{OncoSimXModel} instance.
}
\description{
OncoSimX Model Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{Type}}{Object type (used for \code{print()} method).}

\item{\code{ModelDigest}}{Model digest.}

\item{\code{ModelName}}{Model name.}

\item{\code{ModelVersion}}{Model version.}

\item{\code{ModelMetadata}}{Model metadata.}

\item{\code{ParamsInfo}}{Input parameter information.}

\item{\code{TablesInfo}}{Output table information.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ModelWorksets}}{Data frame of worksets.}

\item{\code{ModelScenarios}}{Data frame of scenarios.}

\item{\code{ModelRuns}}{Data frame of model runs.}

\item{\code{ModelTasks}}{Data frame of model tasks.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXModel-new}{\code{OncoSimXModel$new()}}
\item \href{#method-OncoSimXModel-print}{\code{OncoSimXModel$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModel-new}{}}}
\subsection{Method \code{new()}}{
Create a new OncoSimXModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModel$new(model)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model digest or name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{OncoSimXModel} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModel-print"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModel-print}{}}}
\subsection{Method \code{print()}}{
Print a OncoSimXModel object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModel$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not currently used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
}
