% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModelRun.R
\name{load_model_run}
\alias{load_model_run}
\alias{OncoSimXModelRun}
\title{OncoSimX Model Run Class}
\usage{
load_model_run(model, run)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}
}
\description{
OncoSimX Model Run Class
}
\section{Super class}{
\code{\link[oncosimx:OncoSimXModel]{oncosimx::OncoSimXModel}} -> \code{OncoSimXModelRun}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{RunName}}{Run name.}

\item{\code{RunDigest}}{Run digest.}

\item{\code{RunStatus}}{Run status.}

\item{\code{RunMetadata}}{Run metadata.}

\item{\code{Type}}{Object type (used for \code{print()}).}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXModelRun-new}{\code{OncoSimXModelRun$new()}}
\item \href{#method-OncoSimXModelRun-populate_tables}{\code{OncoSimXModelRun$populate_tables()}}
\item \href{#method-OncoSimXModelRun-get_table}{\code{OncoSimXModelRun$get_table()}}
\item \href{#method-OncoSimXModelRun-populate_params}{\code{OncoSimXModelRun$populate_params()}}
\item \href{#method-OncoSimXModelRun-get_param}{\code{OncoSimXModelRun$get_param()}}
\item \href{#method-OncoSimXModelRun-print}{\code{OncoSimXModelRun$print()}}
\item \href{#method-OncoSimXModelRun-clone}{\code{OncoSimXModelRun$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-new}{}}}
\subsection{Method \code{new()}}{
Create a new OncoSimXModelRun object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$new(model, run)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model digest or name.}

\item{\code{run}}{Run digest, run stamp, or run name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{OncoSimXModel} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-populate_tables"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-populate_tables}{}}}
\subsection{Method \code{populate_tables()}}{
Load all tables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$populate_tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-get_table}{}}}
\subsection{Method \code{get_table()}}{
Retrieve a table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$get_table(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-populate_params"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-populate_params}{}}}
\subsection{Method \code{populate_params()}}{
Load all parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$populate_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-get_param"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-get_param}{}}}
\subsection{Method \code{get_param()}}{
Retrieve a parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$get_param(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Parameter name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-print"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-print}{}}}
\subsection{Method \code{print()}}{
Print a OncoSimXModelRun object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not currently used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
