% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModelRun.R
\name{load_model_run}
\alias{load_model_run}
\alias{load_run}
\alias{OncoSimXModelRun}
\title{OncoSimX ModelRun Class}
\usage{
load_model_run(model, run)

load_run(model, run)
}
\arguments{
\item{model}{Model digest or model name.}

\item{run}{Model run digest, run stamp or run name, modeling task run
stamp or task run name.}
}
\value{
An \code{OncoSimXModelRun} instance.
}
\description{
OncoSimX ModelRun Class
}
\details{
\code{load_run()} is an alias for \code{load_model_run()}.
}
\section{Super class}{
\code{\link[oncosimx:OncoSimXModel]{oncosimx::OncoSimXModel}} -> \code{OncoSimXModelRun}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{RunName}}{Run name.}

\item{\code{RunDigest}}{Run digest.}

\item{\code{RunStamp}}{Run stamp.}

\item{\code{RunMetadata}}{Run metadata.}

\item{\code{Parameters}}{Run parameters.}

\item{\code{Tables}}{Run tables}

\item{\code{Type}}{Object type (used for \code{print()}).}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{RunStatusInfo}}{Run status information.}

\item{\code{RunStatus}}{Run status.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXModelRun-new}{\code{OncoSimXModelRun$new()}}
\item \href{#method-OncoSimXModelRun-print}{\code{OncoSimXModelRun$print()}}
\item \href{#method-OncoSimXModelRun-get_table}{\code{OncoSimXModelRun$get_table()}}
\item \href{#method-OncoSimXModelRun-write_table}{\code{OncoSimXModelRun$write_table()}}
\item \href{#method-OncoSimXModelRun-write_tables}{\code{OncoSimXModelRun$write_tables()}}
\item \href{#method-OncoSimXModelRun-get_log}{\code{OncoSimXModelRun$get_log()}}
\item \href{#method-OncoSimXModelRun-write_log}{\code{OncoSimXModelRun$write_log()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-new}{}}}
\subsection{Method \code{new()}}{
Create a new OncoSimXModelRun object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$new(model, run)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model digest or name.}

\item{\code{run}}{Run digest, run stamp, or run name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{OncoSimXModelRun} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-print"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{OncoSimXModelRun} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not currently used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-get_table}{}}}
\subsection{Method \code{get_table()}}{
Retrieve a table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$get_table(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-write_table"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-write_table}{}}}
\subsection{Method \code{write_table()}}{
Write an output table to disk (CSV).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$write_table(name, file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name.}

\item{\code{file}}{File path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-write_tables"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-write_tables}{}}}
\subsection{Method \code{write_tables()}}{
Write all output tables to disk (CSV).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$write_tables(dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Directory path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-get_log"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-get_log}{}}}
\subsection{Method \code{get_log()}}{
Get console log for model run.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$get_log()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRun-write_log"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRun-write_log}{}}}
\subsection{Method \code{write_log()}}{
Write console log for model run to disk.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRun$write_log(dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Directory to save log file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
}
