% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModelRunSet.R
\name{load_model_runs}
\alias{load_model_runs}
\alias{load_runs}
\alias{OncoSimXModelRunSet}
\title{OncoSimX ModelRunSet  Class}
\usage{
load_model_runs(model, runs)

load_runs(model, runs)
}
\arguments{
\item{model}{Model name or digest.}

\item{runs}{Character vector of model run names, digests, or stamps.}
}
\value{
An \code{OncoSimXModelRunSet} instance.
}
\description{
OncoSimX ModelRunSet  Class
}
\details{
\code{load_runs()} is an alias for \code{load_model_runs()}.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ModelDigest}}{Model digest.}

\item{\code{ModelName}}{Model name.}

\item{\code{ModelVersion}}{Model version.}

\item{\code{Type}}{Object type (used for \code{print()}).}

\item{\code{Params}}{List of input parameters.}

\item{\code{Tables}}{List of output tables.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{RunNames}}{Run names.}

\item{\code{RunDigests}}{Run digests.}

\item{\code{RunStatuses}}{Run statuses.}

\item{\code{RunMetadatas}}{Run metadatas.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXModelRunSet-new}{\code{OncoSimXModelRunSet$new()}}
\item \href{#method-OncoSimXModelRunSet-print}{\code{OncoSimXModelRunSet$print()}}
\item \href{#method-OncoSimXModelRunSet-load_table}{\code{OncoSimXModelRunSet$load_table()}}
\item \href{#method-OncoSimXModelRunSet-load_tables}{\code{OncoSimXModelRunSet$load_tables()}}
\item \href{#method-OncoSimXModelRunSet-get_table}{\code{OncoSimXModelRunSet$get_table()}}
\item \href{#method-OncoSimXModelRunSet-extract_table}{\code{OncoSimXModelRunSet$extract_table()}}
\item \href{#method-OncoSimXModelRunSet-load_param}{\code{OncoSimXModelRunSet$load_param()}}
\item \href{#method-OncoSimXModelRunSet-get_param}{\code{OncoSimXModelRunSet$get_param()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRunSet-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRunSet-new}{}}}
\subsection{Method \code{new()}}{
Create a new OncoSimXModelRunSet object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRunSet$new(model, runs)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model digest or name.}

\item{\code{runs}}{Run digests, run stamps, or run names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{OncoSimXModelSet} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRunSet-print"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRunSet-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{OncoSimXModelRunSet} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRunSet$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not currently used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRunSet-load_table"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRunSet-load_table}{}}}
\subsection{Method \code{load_table()}}{
Load a table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRunSet$load_table(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRunSet-load_tables"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRunSet-load_tables}{}}}
\subsection{Method \code{load_tables()}}{
Load all tables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRunSet$load_tables()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRunSet-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRunSet-get_table}{}}}
\subsection{Method \code{get_table()}}{
Retrieve a table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRunSet$get_table(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRunSet-extract_table"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRunSet-extract_table}{}}}
\subsection{Method \code{extract_table()}}{
Write an output table to disk (CSV).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRunSet$extract_table(name, file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Table name.}

\item{\code{file}}{File path.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRunSet-load_param"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRunSet-load_param}{}}}
\subsection{Method \code{load_param()}}{
Load a parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRunSet$load_param(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Parameter name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelRunSet-get_param"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelRunSet-get_param}{}}}
\subsection{Method \code{get_param()}}{
Retrieve a parameter.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelRunSet$get_param(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Parameter name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble}.
}
}
}
