% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModelTask.R, R/R6ClassWorkset.R
\name{OncoSimXModelTask}
\alias{OncoSimXModelTask}
\alias{load_workset}
\alias{OncoSimXWorkset}
\title{OncoSimX Workset Class}
\usage{
load_workset(model, set, ...)
}
\arguments{
\item{model}{Model digest or model name.}

\item{set}{Name of workset (input set of model parameters).}

\item{...}{Not currently used.}
}
\description{
OncoSimX Workset Class
}
\section{Super classes}{
\code{\link[oncosim:OncoSimXModel]{oncosim::OncoSimXModel}} -> \code{\link[oncosim:OncoSimXModelRun]{oncosim::OncoSimXModelRun}} -> \code{OncoSimXModelTask}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXModelTask-new}{\code{OncoSimXModelTask$new()}}
\item \href{#method-OncoSimXModelTask-clone}{\code{OncoSimXModelTask$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="oncosim" data-topic="OncoSimXModel" data-id="print"><a href='../../oncosim/html/OncoSimXModel.html#method-OncoSimXModel-print'><code>oncosim::OncoSimXModel$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelTask-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelTask$new(model, task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelTask-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelTask-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelTask$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
\section{Super class}{
\code{\link[oncosim:OncoSimXModel]{oncosim::OncoSimXModel}} -> \code{OncoSimXWorkset}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXWorkset-new}{\code{OncoSimXWorkset$new()}}
\item \href{#method-OncoSimXWorkset-clone}{\code{OncoSimXWorkset$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="oncosim" data-topic="OncoSimXModel" data-id="print"><a href='../../oncosim/html/OncoSimXModel.html#method-OncoSimXModel-print'><code>oncosim::OncoSimXModel$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$new(model, set)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
