% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6ClassModelTask.R, R/R6ClassWorkset.R
\name{OncoSimXModelTask}
\alias{OncoSimXModelTask}
\alias{load_workset}
\alias{load_scenario}
\alias{OncoSimXWorkset}
\title{OncoSimX Workset Class}
\usage{
load_workset(model, set)

load_scenario(model, set)
}
\arguments{
\item{model}{Model digest or model name.}

\item{set}{Name of workset (input set of model parameters).}
}
\value{
An \code{OncoSimXWorkset} instance.
}
\description{
OncoSimX Workset Class
}
\details{
\code{load_scenario()} is an alias for \code{load_workset()}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXModelTask-new}{\code{OncoSimXModelTask$new()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXModelTask-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXModelTask-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{OncoSimXModelTask} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXModelTask$new(model, task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model digest or name.}

\item{\code{task}}{Task name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{OncoSimXModelTask} object.
}
}
}
\section{Super class}{
\code{\link[oncosimx:OncoSimXModel]{oncosimx::OncoSimXModel}} -> \code{OncoSimXWorkset}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{WorksetName}}{Workset name.}

\item{\code{WorksetMetadata}}{Workset metadata.}

\item{\code{WorksetDir}}{Workset directory.}

\item{\code{BaseRunDigest}}{Base run digest for input parameters.}

\item{\code{Type}}{Object type (used for \code{print()}).}

\item{\code{Params}}{List of input parameters.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{ReadOnly}}{Workset read-only status.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OncoSimXWorkset-new}{\code{OncoSimXWorkset$new()}}
\item \href{#method-OncoSimXWorkset-get_param}{\code{OncoSimXWorkset$get_param()}}
\item \href{#method-OncoSimXWorkset-refresh_params}{\code{OncoSimXWorkset$refresh_params()}}
\item \href{#method-OncoSimXWorkset-print}{\code{OncoSimXWorkset$print()}}
\item \href{#method-OncoSimXWorkset-copy_params}{\code{OncoSimXWorkset$copy_params()}}
\item \href{#method-OncoSimXWorkset-copy_params_all}{\code{OncoSimXWorkset$copy_params_all()}}
\item \href{#method-OncoSimXWorkset-create_dir}{\code{OncoSimXWorkset$create_dir()}}
\item \href{#method-OncoSimXWorkset-extract_param}{\code{OncoSimXWorkset$extract_param()}}
\item \href{#method-OncoSimXWorkset-extract_params}{\code{OncoSimXWorkset$extract_params()}}
\item \href{#method-OncoSimXWorkset-upload_params}{\code{OncoSimXWorkset$upload_params()}}
\item \href{#method-OncoSimXWorkset-run}{\code{OncoSimXWorkset$run()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-new"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-new}{}}}
\subsection{Method \code{new()}}{
Create a new OncoSimXWorkset object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$new(model, set)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Model digest or name.}

\item{\code{set}}{Workset name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{OncoSimXWorkset} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-get_param"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-get_param}{}}}
\subsection{Method \code{get_param()}}{
Retrieve a parameter table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$get_param(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Parameter name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{tibble}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-refresh_params"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-refresh_params}{}}}
\subsection{Method \code{refresh_params()}}{
Refresh parameters.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$refresh_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-print"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-print}{}}}
\subsection{Method \code{print()}}{
Print a \code{OncoSimXWorkset} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Not currently used.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-copy_params"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-copy_params}{}}}
\subsection{Method \code{copy_params()}}{
Copy a parameter from a base scenario.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$copy_params(names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{names}}{Character vector of parameter names.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-copy_params_all"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-copy_params_all}{}}}
\subsection{Method \code{copy_params_all()}}{
Copy all parameters from a base scenario.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$copy_params_all()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-create_dir"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-create_dir}{}}}
\subsection{Method \code{create_dir()}}{
Create directory for scenario.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$create_dir(dir = ".")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dir}}{Root directory for scenario.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-extract_param"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-extract_param}{}}}
\subsection{Method \code{extract_param()}}{
Write a parameter to disk (CSV format).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$extract_param(name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Parameter name.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-extract_params"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-extract_params}{}}}
\subsection{Method \code{extract_params()}}{
Write all parameters to disk (CSV format).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$extract_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-upload_params"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-upload_params}{}}}
\subsection{Method \code{upload_params()}}{
Upload parameters from a directory.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$upload_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Self, invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OncoSimXWorkset-run"></a>}}
\if{latex}{\out{\hypertarget{method-OncoSimXWorkset-run}{}}}
\subsection{Method \code{run()}}{
Run scenario
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OncoSimXWorkset$run(name, opts = opt_runs(), wait = FALSE, wait_time = 0.1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Run name.}

\item{\code{opts}}{Run options.}

\item{\code{wait}}{Logical. Should we wait until the model run is done?}

\item{\code{wait_time}}{Number of seconds to wait between status checks.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Self, invisibly.
}
}
}
