% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncosim-package.R
\docType{package}
\name{oncosimx-package}
\alias{oncosimx}
\alias{oncosimx-package}
\title{oncosimx: Programmatic interface to the OncoSimX Web Platform}
\description{
This package provides a programmatic interface to the OncoSimX web-based platform. OncoSim is a free, web-based simulation tool that can be used to evaluate cancer control strategies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oncology-outcomes/oncosimx}
  \item \url{https://oncology-outcomes.github.io/oncosimx/}
  \item Report bugs at \url{https://github.com/oncology-outcomes/oncosimx/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{matthew.warkentin@ucalgary.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID})

}
\keyword{internal}
