% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmpp-package.R
\docType{package}
\name{openmpp-package}
\alias{openmpp}
\alias{openmpp-package}
\title{openmpp: Programmatic interface to OpenM++}
\description{
This package provides a programmatic interface to the OpenM++ microsimulation platform.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oncology-outcomes/openmpp}
  \item \url{https://oncology-outcomes.github.io/openmpp/}
  \item Report bugs at \url{https://github.com/oncology-outcomes/openmpp/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{matthew.warkentin@ucalgary.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID})

}
\keyword{internal}
