% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmpp-package.R
\docType{package}
\name{openmpp-package}
\alias{openmpp}
\alias{openmpp-package}
\title{openmpp: Programmatic Interface to OpenM++}
\description{
A programmatic interface to the OpenM++ microsimulation platform. The primary goal of this package is to wrap the OpenM++ Web Service (OMS) to provide OpenM++ users a programmatic interface for the R language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mattwarkentin/openmpp}
  \item \url{https://mattwarkentin.github.io/openmpp/}
  \item Report bugs at \url{https://github.com/mattwarkentin/openmpp/issues}
}

}
\author{
\strong{Maintainer}: Matthew T. Warkentin \email{matthew.warkentin@ucalgary.ca} (\href{https://orcid.org/0000-0001-8730-3511}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item John M. Hutchinson \email{john.hutchinson@ucalgary.ca} [contributor]
}

}
\keyword{internal}
