% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunOptions.R
\name{opts_run}
\alias{opts_run}
\alias{print.OpenMppRunOpts}
\title{Run Configuration Options}
\usage{
opts_run(
  SimulationCases = "5000",
  SubValues = "12",
  RunStamp = NULL,
  Tables = list(),
  ...
)

\method{print}{OpenMppRunOpts}(x, ...)
}
\arguments{
\item{SimulationCases}{Number of cases to simulate. Default is 5000.}

\item{SubValues}{Number of sub values. Default is 12.}

\item{RunStamp}{Run stamp.}

\item{Tables}{List of tables to output.}

\item{...}{Not currently used.}

\item{x}{Object to print.}
}
\description{
Run Configuration Options
}
\details{
The default number of \code{SimulationCases} is low to enable rapid
iteration but should be increased when running a model where the results
are expected to be robust.
}
